Revert Changes from 5cd8acc
The Twitter auto-spam filter has been known to reject similar messages and messages coming in to often, game changes needs a form of cooldown to prevent this issue.  This is evident with a "bug" reported in the forums where the issue was even after a long cooldown the stream-startup message was considered spam and not being posted to Twitter.